+++ src/components/index.ts
@@ -2,4 +2,4 @@ export * from './NavLink';
 export * from './Link';
 export * from './Navigate';
 export * from './Route';
-export * from './RouterSwitch';
+export * from './RouterRoutes';