+++ src/hooks/useRoutesActive.ts
@@ -28,22 +28,13 @@ import { ChildRouteMap } from '../routeFn';
  *                  for each indicating wether or not the route is active
  */
 export const useRoutesActive = <CRM extends ChildRouteMap<any>>(
-  routes: CRM,
-  options?: {
-    strict?: boolean;
-    exact?: boolean;
-  }
+  routes: CRM
 ): { [K in keyof CRM]: boolean } => {
   const { pathname } = useLocation();
 
   return Object.entries(routes)
     .map(v => ({
-      [v[0]]:
-        matchPath(pathname, {
-          path: v[1].fullTemplate,
-          strict: options?.strict ?? v[1].strict,
-          exact: options?.exact ?? v[1].exact,
-        }) != null,
+      [v[0]]: matchPath(pathname, v[1].fullTemplate) != null,
     }))
     .reduce((v, c) => Object.assign(v, c)) as any;
 };