+++ src/routeArgs.ts
@@ -74,27 +74,6 @@ type RouteFnBaseArgs<RO extends RouteOptions> = {
   includeChildren?: boolean;
 
   /**
-   * When true, will only match if the path matches the location.pathname exactly.
-   *
-   * @remark
-   * Taken from https://reactrouter.com/web/api/Route/exact-bool
-   *
-   * @defaultValue true
-   */
-  exact?: boolean;
-
-  /**
-   * When true, a path that has a trailing slash will only match a location.pathname
-   * with a trailing slash. This has no effect when there are additional URL segments in the location.pathname.
-   *
-   * @remark
-   * Taken from https://reactrouter.com/web/api/Route/strict-bool
-   *
-   * @defaultValue false
-   */
-  strict?: boolean;
-
-  /**
    * When true, will match if the path is case sensitive.
    *
    * @remark
@@ -102,5 +81,5 @@ type RouteFnBaseArgs<RO extends RouteOptions> = {
    *
    * @defaultValue false
    */
-  sensitive?: boolean;
+  caseSensitive?: boolean;
 };