+++ src/routeFn.tsx
@@ -54,7 +54,7 @@ export function routeFn<
   args: RouteFnArgs<T, TPM, RO>,
   children?: ChildrenRouterFn<RO, CRM>
 ): RouteNode<T, TPM, CRM, RO> {
-  var params: TemplateParserMap<T>;
+  var params: TemplateParserMap<T> | undefined;
   if (isRouteArgsWithParams(args)) {
     params = args.params;
   }
@@ -185,9 +185,7 @@ export function routeFn<
     template: parsedRoute.pathTemplate,
     children: _children ?? ({} as CRM),
     options: options,
-    exact: args.exact ?? true,
-    strict: args.strict ?? false,
-    sensitive: args.sensitive ?? false,
+    caseSensitive: args.caseSensitive ?? false,
     includeChildren: args.includeChildren ?? true,
     render: middleware
       ? () => middleware!(args.component)