+++ src/types.ts
@@ -3,7 +3,7 @@ import { ParamParser } from "./paramParser";
 /**
  * The React Component for a Route
  */
-export type RouteComponent = React.ComponentType<any>;
+export type RouteComponent = React.ReactElement;
 
 export type RouteOptions = Record<string, any> | undefined;